Consolidated server to one port
Changed server.js to either send the static files for the website (for requests to the root) or send job data (for requests to /jobdata), removing the use of httpserver. Updated run.bat, index.html, and readme.md according to the change. Also updated the gitignore policy to exclude files named test.